refactor metric utility functions to utils.py
