+++ scripts/utils.py
@@ -3,7 +3,166 @@ from pathlib import Path
 from typing import Annotated
 import numpy as np
 import polars as pl
+
+from proteingym.base.dataset import Subsets, Dataset, SEQUENCE
 import typer
+
+
+def get_fold_indices(subsets: Subsets, split: str) -> list[int]:
+    """Get all fold indices for a given split strategy.
+
+    Args:
+        subsets: The Subsets object containing split information.
+        split: The name of the split strategy (e.g., 'random', 'kfold_random').
+
+    Returns:
+        List of fold indices available for the split.
+    """
+    return list(range(len(subsets.slices[split])))
+
+
+def prepare_and_validate_scoring_df(
+    ground_truth: Subsets | Dataset,
+    predicted: Dataset,
+    target: str,
+    split: str | None = None,
+    fold: int | list[int] | None = None,
+) -> pl.DataFrame:
+    """Prepare and validate a scoring dataframe from ground truth and predictions.
+
+    Joins ground truth and predicted datasets on sequence and assay variables,
+    ensuring complete prediction coverage. The returned DataFrame contains both
+    ground truth and predicted values for the specified target, aligned by
+    sequence and variables.
+
+    Args:
+        ground_truth: The ground truth data, either as a complete Dataset or
+            a Subsets object containing dataset slices.
+        predicted: The predicted Dataset containing model predictions for the
+            target. Must have the same structure (assays and variables) as the
+            ground truth.
+        target: The name of the target variable to score (e.g., 'fitness',
+            'binding_affinity'). Must be present in both datasets' assay_targets.
+        split: Required when ground_truth is a Subsets object. The name of the
+            splitting strategy to evaluate (e.g., 'random', 'kfold_random').
+        fold: Required when ground_truth is a Subsets object. Can be:
+            - A single fold index (int) to score one fold
+            - A list of fold indices to score multiple folds in aggregate
+
+    Returns:
+        A Polars DataFrame with columns:
+            - 'sequence': The protein sequence
+            - assay variable columns (e.g., 'temperature', 'pH')
+            - target column: ground truth values
+            - target_pred column: predicted values (with '_pred' suffix)
+
+    Raises:
+        TypeError: If ground_truth is neither a Dataset nor a Subsets object.
+        ValueError: If split or fold is None when ground_truth is a Subsets object.
+        ValueError: If any ground truth records lack corresponding predictions
+            (incomplete coverage).
+
+    Examples:
+        >>> # Score a complete dataset
+        >>> df = prepare_and_validate_scoring_df(
+        ...     ground_truth=test_dataset,
+        ...     predicted=predictions_dataset,
+        ...     target='fitness'
+        ... )
+        >>> # Score a specific fold from dataset slices
+        >>> df = prepare_and_validate_scoring_df(
+        ...     ground_truth=cv_subsets,
+        ...     predicted=predictions_dataset,
+        ...     target='fitness',
+        ...     split='test',
+        ...     fold=0
+        ... )
+    """
+    if isinstance(ground_truth, Dataset):
+        gt_df = ground_truth.to_df(target_names=target)
+        pred_df = predicted.to_df(target_names=target)
+    elif isinstance(ground_truth, Subsets):
+        if split is None or fold is None:
+            raise ValueError(
+                "Both 'split' and 'fold' must be provided when scoring Subsets."
+            )
+
+        fold_indices = [fold] if isinstance(fold, int) else fold
+
+        gt_dfs = []
+        pred_dfs = []
+        for fold_idx in fold_indices:
+            dataset_slice = ground_truth.slices[split][fold_idx]
+            gt_dfs.append(
+                ground_truth.dataset[dataset_slice].to_df(target_names=target)
+            )
+            pred_dfs.append(predicted[dataset_slice].to_df(target_names=target))
+
+        gt_df = pl.concat(gt_dfs, how="vertical_relaxed")
+        pred_df = pl.concat(pred_dfs, how="vertical_relaxed")
+    else:
+        raise TypeError("'ground_truth' must be a Dataset or a Subsets object.")
+
+    # Join on (sequence, variables) to align predictions with ground truth
+    # Only use variables that are actually present and non-null in both DataFrames
+    if isinstance(ground_truth, Subsets):
+        declared_variable_names = [v.name for v in ground_truth.dataset.assay_variables]
+    else:
+        declared_variable_names = [v.name for v in ground_truth.assay_variables]
+
+    variable_names = [
+        var
+        for var in declared_variable_names
+        if var in gt_df.columns
+        and var in pred_df.columns
+        and not gt_df[var].is_null().all()
+        and not pred_df[var].is_null().all()
+    ]
+    join_keys = [SEQUENCE] + variable_names
+
+    joined = gt_df.join(pred_df, on=join_keys, how="inner", suffix="_pred")
+
+    missing_predictions = len(gt_df) - len(joined)
+    if missing_predictions > 0:
+        raise ValueError(f"Missing {missing_predictions} prediction(s).")
+
+    return joined
+
+
+def _get_top_k_from_slice(
+    ground_truth: Subsets | Dataset,
+    split: str | None,
+    fold: int | None,
+) -> int | None:
+    """Extract top_k from slice metadata if available.
+
+    Args:
+        ground_truth: The ground truth data (Dataset or Subsets).
+        split: The split name (required for Subsets).
+        fold: The fold index (required for Subsets, must be int not list).
+
+    Returns:
+        The top_k value from metadata, or None if:
+            - ground_truth is not a Subsets object
+            - split or fold is not provided
+            - fold is a list (not supported for recovery)
+            - metadata doesn't exist or doesn't contain top_k
+    """
+    if not isinstance(ground_truth, Subsets):
+        return None
+    if split is None or fold is None:
+        return None
+    if isinstance(fold, list):
+        return None
+
+    try:
+        dataset_slice = ground_truth.slices[split][fold]
+        if dataset_slice.metadata is None:
+            return None
+        top_k = dataset_slice.metadata.get("top_k")
+        return int(top_k) if top_k is not None else None
+    except (KeyError, IndexError, AttributeError, ValueError, TypeError):
+        return None
 
 
 def aggregate_metrics(
